sexpression parser: improve and generalize ignored syntax
Make "ignorable" cases extensible; in particular, the s-expression
comment prefix is now part of the extensible system, and is also
recognized anywhere within an s-expression.  For example "(a #;q)"
correctly reads as (a).  This was the last difficult change in the
reader.  What remains is a set of more or less trivial bits, such as
string and character escaping, and multiple radix support for numbers.

* bootstrap/scheme/sexpression-parser.e: Here.