input ports, reader: fix EOF handling
Make #\eof a special "character", defined as io:eof, which is the
constant -1.  Make EOF-checking on generic input port, file inout
ports, string input ports and readline input ports return a cached
state, which is only updated after first reading a character (which
may now return #\eof); this is consistent with the behavior of fgetc
in the C library.  eof? still works the old way in backtrackable input
ports.

Make #eof a new unique s-expression.  reader:read returns the #eof
s-expression when reading at the end of file.  It is now possible to
reliably find the end of the input.

* bootstrap/scheme/sexpression-parser.e: Here.
* bootstrap/scheme/epsilon1.scm: Here.
* bootstrap/scheme/scratch.scm: Here.
* bootstrap/scheme/core.e: Here.