unexec: also permit to dump a given symbol table
symbol:table is just the default for the old procedure and macro; make
them just wrappers calling generic versions taking the table as a parameter.

e1:unexec and its aliases are not perfect yet, since they rely on
making a closure which then is bound in the primary symbol table.

* bootstrap/scheme/unexec.e: Here.
* bootstrap/scheme/epsilon1.scm: Update e1:unexec and friends.