overhaul compiler
Improve the compiler interface, making it more similar to e1:unexec:
the user can now specify a sequence of main expressions to run in the
compiled program.  The compiler is still only for non-self-modifying
programs, and generates very bad code.

There is still no clean compiler driver, but it's not easy to generate
assembly files.  Add the two ugly scripts COMPILE-NATIVE and COMPILE-C
as temporary compiler drivers, to assemble and link compiler-generated
files into executables.

* bootstrap/scheme/compiler.e: Here.
* driver/driver.c: Here.

* bootstrap/scheme/temporary-c-driver/w.c: Delete.

* TEST-NATIVE: Change into...
* COMPILE-NATIVE: ...This.
* COMPILE-C: Add, following COMPILE-NATIVE.

* Makefile.am: Here.