REPL: add a powerful REPL command definition facility; use it for ,show
Add the macro-defining macro repl:define-command, which permits to
define a REPL command calling a given procdure; the procedure will
take the s-expression following the prefix as its sole argument.

Use the above functionality to define the REPL command ,show.  Its
associated procedure uses io:file-content-as-byte-vector to include
the GNU GPL text from the source tree into the program to be unexeced;
the file will not be used at run time.  Update the REPL banner.

* bootstrap/scheme/repl.e: Here.