movinggc overhaul
A big whole-night overhaul of my semispace exact-pointer-finding
garbage collector including fixes, optimizations, cosmetic changes and
algorithmic improvements.  The collector is not yet integrated into
epsilon, but it will serve as a reasonable C benchmark against which
to compare the future native implementation in epsilon.

This is not the parallel collector.  It's a completely different
design, sequential and comparatively simple but very fast,
particularly with high allocation rate.  It can be adapted into the
young generation of a generational and/or parallel GC.

It might actually make sense to package this collector separately from
epsilon, as other people might find it useful.

Exact pointer finding is very annoying to be used from C, but the
interface should be suitable for compilers and interpreters, thanks to
the idea of asynchronous root scanning.  Write me if you are
interested.

* movinggc/movinggc.c, movinggc/movinggc.h, movinggc/tags.h,
* movinggc/features.h, movinggc/driver.c: Here

* movinggc/CLEAN, movinggc/GO, movinggc/README, movinggc/TODO: New.