tagged runtime: don't shift when tagging foreign pointers
Foreign pointers are now tagged as non-pointers, but without shifting; this is
needed when low-order bits are 1.  Before this change even having the bit on the
right of the most significant one set to 1 caused sign extension while
right-shifting for untagging, which screwed up the resulting pointer.  Now the
only requirement is for the foreign pointer to be aligned to two bytes (checked
on ENABLE_DEBUG).

* runtime/data-tagged.c: Here.
* runtime/c-primitives.c: feof primitive: untag with
epsilon_value_to_foreign_pointer, which now behaves differently from
epsilon_value_to_epsilon_int.