core: implement fixnum square root
The algorithm is not optimal, but not terrible either.  I use Newton-Raphson,
which is guaranteed to converge monotonically (and I exploit monotonicity in the
exit condition) when the initial guess is larger than the result.  This is all
fine, but my first guess is too crude: I should use a count-leading-zeros
instruction.

* bootstrap/scheme/core.e: Add fixnum:sqrt.