epsilon0 C interpreter: undo the previous change
Allocating imperative stacks on the C heap would be the correct solution in an
ideal world, but that interferes with GC roots in Guile 1, which is still the
most convenient version to use for bootstrapping.  Undo the change, keeping the
"good" version in an #if 0 block.

* runtime/epsilon0-interpreter.c: Undo previous change.
* Makefile.am: Increase Guile 2 stack height thru an environment variable.
This will be the correct solution, in theory; I hope I will bootstrap away
from Guile soon enough not to ever need it, but I might forget about the
workaround if not.