bootstrap: speed-up, clean-up and simplify
Bootstrapping from Guile is now performed with *two* unexec's, to make the
process faster and to rebuild build less often.  Only the first unexec is run on
top of guile+whatever: after it, the epsilon image interpreter is used to exec,
load some advanced language features (which should be faster than with Guile)
and to unexec into the final REPL.  More code can and will be moved out of the
first phase, for better performance and with the goal of minimizing remaining
dependencies from Guile.

The REPL is now automatically built as a dependency of the main target.

* bootstrap/scheme/bootstrap.scm: This is the new first phase, running on top of
guile+whatever.
* bootstrap/scheme/unexec-repl.e: This is the new second phase, running on top of
a native epsilon image interpreter.
* Makefile.am: Here.
* bootstrap/scheme/quick-start.scm: Delete: this was confusing and is now unneeded.
* README: Explain how to run the thing in a simple way