core: always call symbol:intern-uninterned! with two args; bootstrapping fixes
Make sure that symbol:intern-uninterned! is always called with two parameters.
In one rarely tested code path I was still calling it with one parameter, which
has been wrong for a long time.

Don't use #| ... |# comments, which are not currently recognized by the epsilon1
reader.

Define symbol:table only if running in Guile; doing it unconditionally would
break bootstrap from epsilon1.

io:write-symbol: print symbols not interned in the primary symbol table in a
different color, so that bootstrapping bugs are easier to spot.