bootstrap: simplify and bootstrap from epsilon1 by default
When bootstrapping a REPL obtained from epsilon1, which is now the default REPL,
do not use the two-step process needed for my Guile implementation, which first
unexecs a REPL generator and only from there the REPL.

Distribute a compressed bootstrap REPL to be used for the initial compilation.
Right now it's still possible to generate a bootstrap REPL from the source only,
using Guile, but this will change.

* configure.ac, Makefile.am: Build system changes
* bootstrap/scheme/bootstrap-from-epsilon1-secondary.e: Bootstrap in one step.