pushover: slight simplification/very minor optimization
po:valid-move-as-string? and po:string->move now expect a valid string to be
trimmed, and I trim the string typed in by the user before calling those
procedures.  This lets me output a nicer error message, showing the input string
trimmed, and is actually *less* expensive than before.  Not that this counts.
My e1:cond with bind clauses is coming in handy.

* examples/pushover.e: Here.