core: define whatever:zero? as a conditional, without using the primitive
The primitive whatever:zero? is usually a bad idea if I want to compile into
efficient code.  Simply defining the wrapper procedure whatever:zero? as a
conditional yields a conditional which is very easy to optimize away in the
common cases (when used within a conditional discriminand), generating zero
assembly instructions.  This is important, as whatever:zero? is also used in the
definition of e1:not.

* bootstrap/scheme/core.e: Here.