(old) compiler: C backend fixes
epsilon_run_thread_context calls a compiler-generated procedure, and was always
supposed to pass it a parameter.  I had forgotten to pass it but, astonishingly,
I had never seen a crash in years on MIPS, amd64 or x86 -- until my first test
on Sparc today.

Second (smaller) bugfix: in the (old) epsilon compiler, link with the C runtime
instead of the native runtime.  This lets compilation work also on targets
without a native backend.

* runtime/backend-specific/c/backend-specific-c.c: Fix bug.

* bootstrap/scheme/compiler.e: Link with the C runtime, not the native runtime.